storage: group more consistency tests
Release note: None
